refactor(kikan): consolidate profile_db_middleware into kikan (Wave A.5)
Audit found no true drift between `services/api/src/profile_db.rs` and
`crates/kikan/src/profile_db.rs` — but the middleware, which the
file-header claimed was welded to `AppState`, in fact only touched
`db_for(mode)` and `active_profile`, both already on `PlatformState`
since the platform-slice lift. Moving the function puts the extractor
and its only producer in one file and eliminates the vertical's need
for a module just to forward kikan surfaces.

Changes:
- Move `profile_db_middleware` from `services/api/src/profile_db.rs`
  (deleted) to `crates/kikan/src/profile_db.rs`, parameterized on
  `State<PlatformState>` instead of `State<SharedState>`.
- Update `services/api::build_app_inner` mount site to
  `from_fn_with_state(state.platform_state(), kikan::profile_db::profile_db_middleware)`.
- Rewrite `profile_db.rs` file header to reflect the new home.

Route shapes, layer ordering (ProfileDb innermost → auth → trace →
security-headers → host-allowlist), and `profile_middleware.feature`
BDD contract all unchanged.

Verification: kikan lib 205 · kikan BDD 49/49 · mokumo-api lib 63 +
integration 145 · clippy clean · fmt clean · I1/I2/I3/I4/I5 green.
Failing mokumo-api BDD scenarios (WebSocket/logo/restore) are
pre-existing on the baseline commit (confirmed via stash/restore).

(#508)
