+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Internal
 
+- **`profile_db_middleware` consolidated into `kikan::profile_db` (Wave A.5)**: Relocates the profile-routing middleware from `services/api/src/profile_db.rs` (deleted) into `crates/kikan/src/profile_db.rs` alongside the `ProfileDb` / `ActiveProfile` extractors. The middleware is parameterized on `State<PlatformState>` — it already only touched `db_for(mode)` and `active_profile`, both present on `PlatformState` since the platform-slice lift, so the prior file-header claim that it "cannot move into platform code without dragging AppState along" was stale. Mount site in `services/api::build_app_inner` updated from `state.clone()` + `profile_db::profile_db_middleware` to `state.platform_state()` + `kikan::profile_db::profile_db_middleware`; route shapes, layer ordering, and `profile_middleware.feature` BDD contract all unchanged. (#508)
 - **UserRepo composite-method atomicity (Wave A.3b)**: Adds two new composite methods on `SeaOrmUserRepo` — `create_user_with_codes(&CreateUser, codes_count)` (generic user + recovery-codes batch with activity-log-in-txn) and `bootstrap_admin_with_codes(email, name, password)` (first-admin bootstrap with an `ALREADY_BOOTSTRAPPED` guard checked *inside* the transaction, so concurrent bootstraps race-safely). The third method named in the feature, `regenerate_codes_with_log`, is satisfied by the existing `regenerate_recovery_codes`. Introduces `kikan::auth::repo::BootstrapError` (`AlreadyBootstrapped`, `Domain(DomainError)`) with `From<BootstrapError> for ControlPlaneError` so handlers render the wire code without string-sniffing. Adds `ActivityAction::Bootstrap` (displays as `"bootstrap"`). Un-tags `user_repo_atomicity.feature` — all 6 scenarios pass (create success, validation rollback via `codes_count` out-of-range, regenerate success, regenerate log-failure rollback via a `RAISE(ABORT)` trigger on `activity_log`, bootstrap success, bootstrap `ALREADY_BOOTSTRAPPED` rejection). Adds `crates/kikan/tests/bdd_world/user_repo_atomicity_steps.rs` and `crates/kikan/tests/user_repo_trait_signature.rs` as a compile-time witness that `UserRepository` leaks no `sea_orm::DatabaseTransaction` or `sea_orm::TransactionTrait` into its public surface. Composite-method transactional atomicity is now enforced at both the structural (trait signature) and behavioral (BDD) levels. (#508)
 - **Activity HTTP handler lift into `kikan::platform::activity_http`**: Relocates `services/api/src/activity/mod.rs` (~55 LOC) to `crates/kikan/src/platform/activity_http/mod.rs`. The `GET /api/activity` list endpoint (optional `entity_type`, `entity_id`, `page`, `per_page` filters) now lives alongside the other platform handlers. Module named `activity_http` to avoid collision with `kikan::activity` (the writer/repo). `activity_router()` is generic over state (`Router<S>`) because the handler uses only per-request extractors (`ProfileDb`, `Query`). services/api mount updated to `kikan::platform::activity_http::activity_router()`; `services/api/src/activity/` deleted. Route shape and pagination semantics unchanged. (#508)
 - **User admin HTTP handler lift into `kikan::platform::users`**: Relocates `services/api/src/user/mod.rs` (~90 LOC) to `crates/kikan/src/platform/users/mod.rs` so the admin-only user mutation endpoints (`DELETE /api/users/{id}` soft delete, `PATCH /api/users/{id}/role` role update) live alongside the other platform handlers. `user_admin_router()` is generic over state (`Router<S>` where `S: Clone + Send + Sync + 'static`) because both handlers rely exclusively on per-request extractors (`AuthSession`, `ProfileDb`) — no singleton dependencies required. services/api mount updated from `user::router()` to `kikan::platform::users::user_admin_router()`. Route shapes, HTTP status codes, and auth/admin-guard semantics unchanged. `services/api/src/user/` deleted. Composite-method atomicity (the `create_user_with_codes`/`regenerate_codes_with_log`/`bootstrap_admin_with_codes` contract pinned by `user_repo_atomicity.feature`) is deferred to a follow-up wave: those methods do not yet exist in `kikan::auth::repo` and the feature stays `@future`-tagged until that work lands. (#508)
